Rename in `src/main`: autoUpdate → launcherUpdate
