+++ src/main/launcherUpdate.ts
